refactor(energy): Move common functions to shared module, and add documentation
